requirements.txt
# Core
python-dotenv
requests

# Web/API
fastapi
uvicorn
jinja2

# Audio/Voice
SpeechRecognition
pyaudio
edge-tts
pygame

# AI/LLM
google-genai
openai

# System/Tools
pyautogui
AppOpener
psutil
schedule
yt-dlp
duckduckgo-search
groq
putergenai
